+++ src/main/java/livenessanalysis/UnusedValue.java
@@ -2,20 +2,20 @@ package livenessanalysis;
 
 import symboltable.Symbol;
 
-/** Data class for a variable that is either never assigned or read. */
-public class UnusedVariable {
+/** Data class for a variable value that is never read. */
+public class UnusedValue {
   private final int lineNumber;
   private final String statementType;
   private final Symbol symbol;
 
   /**
-   * Data class for a variable that is either never assigned or read.
+   * Data class for a variable value that is never read.
    *
    * @param lineNumber    Line number of the declaration or write statement.
    * @param statementType Type of declaration or write statement.
-   * @param symbol        Symbol that is never assigned or read.
+   * @param symbol        Symbol that is never read.
    */
-  public UnusedVariable(int lineNumber, String statementType, Symbol symbol) {
+  public UnusedValue(int lineNumber, String statementType, Symbol symbol) {
     this.lineNumber = lineNumber;
     this.statementType = statementType;
     this.symbol = symbol;