+++ src/test/java/ParsingTest.java
@@ -267,9 +267,16 @@ public class ParsingTest {
   }
 
   @Test
-  public void parseunusedVariableValues() {
+  public void parseUnusedVariableValues() {
     EasyCompiler easyCompiler = new EasyCompiler(
         pathTestFilesLiveness + "unusedVariableValues.easy");
+    assertTrue(easyCompiler.parse());
+  }
+
+  @Test
+  public void parseUnusedDeclarations() {
+    EasyCompiler easyCompiler = new EasyCompiler(
+        pathTestFilesLiveness + "unusedDeclarations.easy");
     assertTrue(easyCompiler.parse());
   }