+++ src/test/java/symboltable/SymbolTableTest.java
@@ -65,4 +65,24 @@ public class SymbolTableTest {
     assertEquals(Type.FLOAT, symbolTable.getSymbolType("scopeB", "symbol"));
     assertEquals(new ArrayList<String>(List.of("scopeA", "scopeB")), symbolTable.getScopeNames());
   }
+
+  @Test
+  public void getNonArgumentSymbols() {
+    SymbolTable symbolTable = new SymbolTable();
+    symbolTable.addNewScope("f", Type.INT);
+    symbolTable.addFunctionArgumentType("f", Type.INT);
+    symbolTable.addSymbolToScope("f", "argumentName", Type.INT);
+    symbolTable.addSymbolToScope("f", "symbolA", Type.INT);
+    symbolTable.addSymbolToScope("f", "symbolB", Type.INT);
+
+    List<Symbol> nonArgumentSymbols = symbolTable.getNonArgumentSymbols("f");
+
+    assertEquals(3, symbolTable.countSymbolsInScope("f"));
+    assertEquals(1, symbolTable.getNumberOfArguments("f"));
+    assertEquals(2, nonArgumentSymbols.size());
+    assertEquals(1, symbolTable.getVariableNumber("f", "symbolA"));
+    assertEquals(2, symbolTable.getVariableNumber("f", "symbolB"));
+    assertEquals(Type.INT, symbolTable.getSymbolType("f", "symbolA"));
+    assertEquals(Type.INT, symbolTable.getSymbolType("f", "symbolB"));
+  }
 }